Move subscription API to separate, independent router
